Added comments, made refactoring
The most of the time I lost by trying to write correct interaction
inside TrainingCardsFragment. So it can look too complicated and
should be refactored again. It was really hard to understand why
things happen and I am not sure that I clearly got it. So the number 1 task
for me for the near future is to deeply understand how this mechanism
works.